requirements.txt
langchain
langchain-google-genai
langchain-postgres
langchain-community
psycopg[binary]
pypdf
python-dotenv